rename Java package from mathias to segware
