+++ android/src/main/java/com/segware/MjpegPlayerView.java
@@ -1,4 +1,4 @@
-package com.mathias;
+package com.segware;
 
 import android.content.Context;
 import android.content.res.Configuration;
@@ -6,10 +6,6 @@ import android.net.Uri;
 import android.util.Log;
 import android.widget.LinearLayout;
 
-import com.facebook.react.bridge.Arguments;
-import com.facebook.react.bridge.ReactContext;
-import com.facebook.react.bridge.WritableMap;
-import com.facebook.react.uimanager.events.RCTEventEmitter;
 import com.mjpeg.DisplayMode;
 import com.mjpeg.Mjpeg;
 import com.mjpeg.MjpegSurfaceView;
