+++ android/src/main/java/com/segware/RNMjpegPlayerModule.java
@@ -1,10 +1,8 @@
 
-package com.mathias;
+package com.segware;
 
 import com.facebook.react.bridge.ReactApplicationContext;
 import com.facebook.react.bridge.ReactContextBaseJavaModule;
-import com.facebook.react.bridge.ReactMethod;
-import com.facebook.react.bridge.Callback;
 
 public class RNMjpegPlayerModule extends ReactContextBaseJavaModule {
